+++ CHANGELOG.md
@@ -36,6 +36,7 @@ const handleDocumentLoad = (e: DocumentLoadEvent) => {
 -   Can't close the menu when using with the [Fluent UI](https://developer.microsoft.com/en-us/fluentui)
 -   The `activateTab()` method provided by the [default layout](https://react-pdf-viewer.dev/plugins/default-layout) plugin shouldn't toggle the tab
 -   The `LocalePopover` component doesn't update the localization
+-   The pages' container overflows on initial render when setting the default scale to `SpecialZoomLevel.PageWidth`
 
 ## v2.7.1